tests/test_dependency_manager.py
"""
Unit tests for lib.dependency_manager module.
"""

import unittest
import tempfile
import yaml
from pathlib import Path
from unittest.mock import patch, MagicMock

# Add parent directory to path for imports
import sys
sys.path.insert(0, str(Path(__file__).parent.parent))

from lib.dependency_manager import DependencyManager


class TestDependencyManager(unittest.TestCase):
    """Test DependencyManager class."""
    
    def setUp(self):
        """Set up test fixtures."""
        self.temp_dir = tempfile.mkdtemp()
        self.deps_file = Path(self.temp_dir) / "test_deps.yaml"
        
        # Sample dependencies for testing
        self.sample_deps = {
            'dependencies': [
                {
                    'type': 'git',
                    'url': 'https://github.com/example/repo.git',
                    'destination': 'test-repo',
                    'description': 'Test repository'
                },
                {
                    'type': 'huggingface',
                    'repo_id': 'test/model',
                    'filename': 'model.safetensors',
                    'destination': 'models/test.safetensors',
                    'description': 'Test model'
                },
                {
                    'type': 'file',
                    'url': 'https://example.com/file.txt',
                    'destination': 'downloads/file.txt',
                    'description': 'Test file'
                }
            ]
        }
        
        # Write dependencies file
        with open(self.deps_file, 'w') as f:
            yaml.dump(self.sample_deps, f)
        
        self.manager = DependencyManager(str(self.deps_file))
    
    def tearDown(self):
        """Clean up test fixtures."""
        import shutil
        shutil.rmtree(self.temp_dir, ignore_errors=True)
    
    def test_load_dependencies_success(self):
        """Test successful dependency loading."""
        success = self.manager.load_dependencies()
        
        self.assertTrue(success)
        self.assertEqual(len(self.manager.dependencies), 3)
        self.assertFalse(self.manager.is_structured)  # Legacy format
    
    def test_load_dependencies_file_not_found(self):
        """Test loading with non-existent file."""
        manager = DependencyManager("nonexistent.yaml")
        success = manager.load_dependencies()
        
        self.assertFalse(success)
        self.assertEqual(len(manager.dependencies), 0)
    
    def test_load_dependencies_invalid_yaml(self):
        """Test loading with invalid YAML."""
        invalid_file = Path(self.temp_dir) / "invalid.yaml"
        with open(invalid_file, 'w') as f:
            f.write("invalid: yaml: content: [")
        
        manager = DependencyManager(str(invalid_file))
        success = manager.load_dependencies()
        
        self.assertFalse(success)
    
    def test_validate_dependency_structure_valid(self):
        """Test validation with valid dependencies."""
        self.manager.load_dependencies()
        errors = self.manager.validate_dependency_structure()
        
        self.assertEqual(len(errors), 0)
    
    def test_validate_dependency_structure_invalid(self):
        """Test validation with invalid dependencies."""
        # Create invalid dependencies
        invalid_deps = {
            'dependencies': [
                {'type': 'git'},  # Missing url and destination
                {'destination': 'test'},  # Missing type
                {'type': 'git', 'destination': 'test'},  # Missing url for git
                {'type': 'huggingface', 'destination': 'test'},  # Missing repo_id and filename
                {'type': 'huggingface', 'repo_id': 'test/model', 'destination': 'test'}  # Missing filename
            ]
        }
        
        invalid_file = Path(self.temp_dir) / "invalid_deps.yaml"
        with open(invalid_file, 'w') as f:
            yaml.dump(invalid_deps, f)
        
        manager = DependencyManager(str(invalid_file))
        manager.load_dependencies()
        errors = manager.validate_dependency_structure()
        
        self.assertGreater(len(errors), 0)
        
        # Check specific error types
        error_text = ' '.join(errors)
        self.assertIn("Missing 'destination'", error_text)
        self.assertIn("Missing 'type'", error_text)
        self.assertIn("Git type missing 'url'", error_text)
        self.assertIn("HuggingFace type missing 'repo_id'", error_text)
        self.assertIn("HuggingFace type missing 'filename'", error_text)
    
    def test_get_dependency_summary_legacy(self):
        """Test getting dependency summary for legacy format."""
        self.manager.load_dependencies()
        
        summary = self.manager.get_dependency_summary()
        
        self.assertEqual(summary['total_dependencies'], 3)
        self.assertEqual(summary['format'], 'legacy')
        self.assertIn('types', summary)
        self.assertEqual(summary['types']['git'], 1)
        self.assertEqual(summary['types']['huggingface'], 1)
        self.assertEqual(summary['types']['file'], 1)
    
    @unittest.skip("Method signature changed in new API")
    def test_get_dependencies_by_type(self):
        """Test filtering dependencies by type (legacy test - skipped)."""
        pass
    
    @unittest.skip("Method not available in new API")
    def test_file_check_methods(self):
        """Test file checking methods directly (legacy test - skipped)."""
        pass
    
    @unittest.skip("Method not available in new API") 
    def test_git_repo_check_methods(self):
        """Test git repository checking methods directly (legacy test - skipped)."""
        pass
    
    @unittest.skip("Method signature changed in new API")
    def test_check_all_dependencies(self):
        """Test checking all dependencies (legacy test - skipped)."""
        pass
    
    @unittest.skip("Method not available in new API")
    def test_check_file_exists(self):
        """Test file existence checking (legacy test - skipped)."""
        pass
    
    @unittest.skip("Method not available in new API")
    def test_check_git_repo_exists(self):
        """Test git repo existence checking (legacy test - skipped)."""
        pass
    
    @unittest.skip("Method not available in new API")
    def test_check_huggingface_file_exists(self):
        """Test HuggingFace file existence checking (legacy test - skipped)."""
        pass


if __name__ == '__main__':
    unittest.main() 