requirements-dev.txt
PyYAML==6.0.1
requests==2.31.0
tqdm 